[FrameworkBundle][Serializer] Move SerializerPass to the Serializer
