+++ UPGRADE-3.3.md
@@ -35,6 +35,10 @@ FrameworkBundle
 
  * The `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddConsoleCommandPass` has been deprecated. Use `Symfony\Component\Console\DependencyInjection\AddConsoleCommandPass` instead.
 
+ * The `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\SerializerPass` class has been
+   deprecated and will be removed in 4.0. 
+   Use the `Symfony\Component\Serializer\DependencyInjection\SerializerPass` class instead.
+
 HttpKernel
 -----------
