+++ UPGRADE-4.0.md
@@ -156,6 +156,9 @@ FrameworkBundle
 
  * The `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddConsoleCommandPass` has been removed. Use `Symfony\Component\Console\DependencyInjection\AddConsoleCommandPass` instead.
 
+ * The `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\SerializerPass` class has been removed. 
+   Use the `Symfony\Component\Serializer\DependencyInjection\SerializerPass` class instead.
+
 SecurityBundle
 --------------
